+++ app/core/path-manager.js
@@ -89,14 +89,14 @@ module.exports = {
    * @returns {boolean}
    */
   isRanFromPackagedNode() {
-    return __filename.indexOf('/snapshot/') > -1;
+    return __filename.indexOf('/snapshot/') > -1 || __filename.indexOf('C:\\snapshot\\') > -1;
   },
 
   /**
    * @returns {boolean}
    */
   isRanAsStandaloneNodeProject() {
-    return process.argv[1].indexOf('bin/admin') > -1;
+    return process.argv[1].indexOf('bin/admin') > -1 || process.argv[1].indexOf('bin\\admin') > -1;
   },
 
   /**