+++ app/core/server-process-manager.js
@@ -1,16 +1,17 @@
 /**
  * server-process-manager.js
  */
-const { exec, execSync } = require('child_process');
-const serverDataService  = require('./eqemu-data-service-client.js');
-const pathManager        = require('./path-manager');
-const fs                 = require('fs');
-const path               = require('path')
-const os                 = require('os')
-const util               = require('util')
-const psList             = require('ps-list');
-const debug              = require('debug')('eqemu-admin:process-manager');
-const config             = require('./eqemu-config-service')
+const {exec, execSync}  = require('child_process');
+const serverDataService = require('./eqemu-data-service-client.js');
+const pathManager       = require('./path-manager');
+const fs                = require('fs');
+const path              = require('path');
+const util              = require('util');
+const psList            = require('ps-list');
+const debug             = require('debug')('eqemu-admin:process-manager');
+const config            = require('./eqemu-config-service');
+const systemProc        = require('process');
+const os                = use('/app/core/operating-system');
 
 /**
  * Constants
@@ -38,11 +39,14 @@ module.exports = {
   /**
    * Launcher initialization
    * @param options
+   * @param skipConfigWatch
    */
-  init: function (options) {
+  init: function (options, skipConfigWatch = false) {
     this.launchOptions = options;
 
-    config.init();
+    process.title = "eqemu-server-launcher";
+
+    config.init(skipConfigWatch);
 
     let self = this;
     this.serverProcessNames.forEach(function (process_name) {
@@ -56,22 +60,24 @@ module.exports = {
    * @returns {number}
    */
   getWatchDogPollDelta() {
-    return (Math.floor(new Date() / 1000) - this.lastProcessPollTime)
+    return (Math.floor(new Date() / 1000) - this.lastProcessPollTime);
   },
 
   /**
    * @returns {boolean}
    */
   isEqemuServerOnline() {
-    let isServerOnline = false
-    let self = this
+    let isServerOnline = false;
+    let self           = this;
     this.serverProcessNames.forEach(function (processName) {
       if (self.processCount[processName] > 0) {
-        isServerOnline = true
+        isServerOnline = true;
       }
     });
 
-    return isServerOnline
+    debug('isEqemuServerOnline [%s]', isServerOnline);
+
+    return isServerOnline;
   },
 
   /**
@@ -79,18 +85,32 @@ module.exports = {
    * @returns {Promise<void>}
    */
   start: async function (options = []) {
-    this.init(options);
+    this.init(options, true);
 
     await this.pollProcessList();
 
-    debug('server is [%s]', (this.isEqemuServerOnline() ? "online" : "offline"))
+    if (await this.isLauncherBooted()) {
+      debug('Launcher is already booted... exiting...');
+      console.log('Launcher is already booted... exiting...');
+      process.exit();
+    }
+
+    debug('server is [%s]', (this.isEqemuServerOnline() ? 'online' : 'offline'));
 
     /**
      * Shared memory
      */
-    if (config.getAdminPanelConfig('launcher.runSharedMemory', true) && !this.isEqemuServerOnline()) {
-      debug("Running shared memory");
-      execSync('./bin/shared_memory', { cwd: pathManager.emuServerPath }).toString();
+    if (config.getAdminPanelConfig('launcher.runSharedMemory') && !this.isEqemuServerOnline()) {
+      debug('Running shared memory');
+
+      if (os.isLinux()) {
+        execSync('./bin/shared_memory', {cwd: pathManager.emuServerPath}).toString();
+      }
+
+      if (os.isWindows()) {
+        execSync('bin\\shared_memory', {cwd: pathManager.emuServerPath}).toString();
+      }
+
     }
 
     if (config.getAdminPanelConfig('launcher.runLoginserver', false)) {
@@ -105,6 +125,8 @@ module.exports = {
 
     this.startWatchDog();
 
+    config.setAdminPanelConfig('launcher.isRunning', true);
+
     /**
      * Main launcher loop
      */
@@ -118,7 +140,7 @@ module.exports = {
       let self = this;
       for (const process_name of ['loginserver', 'ucs', 'world', 'queryserv']) {
         if (self.doesProcessNeedToBoot(process_name) && !await self.isProcessRunning(process_name)) {
-          debug('starting unique process [' + process_name + ']')
+          debug('starting unique process [' + process_name + ']');
           self.startProcess(process_name);
         }
       }
@@ -143,7 +165,7 @@ module.exports = {
   doesProcessNeedToBoot: function (process_name) {
     if (this.erroredStartsCount[process_name] >= this.erroredStartsMaxHalt) {
       if (!this.hasErroredHaltMessage[process_name]) {
-        console.error('Process [%s] has tried to boot too many (%s) times... Halting attempts', process_name, this.erroredStartsMaxHalt)
+        console.error('Process [%s] has tried to boot too many (%s) times... Halting attempts', process_name, this.erroredStartsMaxHalt);
         this.hasErroredHaltMessage[process_name] = 1;
       }
 
@@ -164,7 +186,7 @@ module.exports = {
       return this.launchOptions && this.launchOptions.withQueryserv && this.processCount[process_name] === 0;
     }
 
-    debug('[doesProcessNeedToBoot] returning [%s]', this.processCount[process_name] === 0)
+    debug('[doesProcessNeedToBoot] [%s] returning [%s]', process_name, this.processCount[process_name] === 0);
 
     return this.processCount[process_name] === 0;
   },
@@ -181,8 +203,10 @@ module.exports = {
    * @returns {Promise<void>}
    */
   stopServer: async function () {
-    this.systemProcessList = await psList();
-    let self               = this;
+    this.init([], true);
+
+    await this.pollProcessList();
+    let self = this;
 
     debug('[stopServer] Stopping server...');
 
@@ -208,6 +232,8 @@ module.exports = {
       }
     });
 
+    config.setAdminPanelConfig('launcher.isRunning', false);
+
     return this;
   },
 
@@ -217,13 +243,53 @@ module.exports = {
   async isLauncherBooted() {
     let isLauncherBooted = false;
     await this.pollProcessList();
-    this.systemProcessList.forEach(function (process) {
-      if (process.cmd.includes('server-launcher')) {
-        isLauncherBooted = true;
+    this.systemProcessList.forEach(function (proc) {
+
+      if (os.isLinux()) {
+
+        /**
+         * 1) Make sure we're looking for a command running the launcher
+         * 2) Make sure it isn't this process
+         * 3) Make sure it's not a while loop that is keeping it alive
+         */
+        if (
+          proc.cmd.includes('server-launcher') &&
+          proc.cmd.includes('admin') &&
+          parseInt(proc.pid) !== parseInt(systemProc.pid) &&
+          !proc.cmd.includes('while')
+        ) {
+          isLauncherBooted = true;
+        }
+      }
+
+      if (os.isWindows()) {
+
+
+        if (
+          proc.cmd.includes('server-launcher') &&
+          parseInt(proc.pid) !== parseInt(systemProc.pid)
+        ) {
+
+          debug("Launcher is already booted");
+          debug(proc);
+          debug("Current system proc [%s]", systemProc.pid);
+          debug("process.pid [%s]", proc.pid);
+
+          isLauncherBooted = true;
+        }
       }
     });
 
+    debug('isLauncherBooted [%s]', isLauncherBooted);
+
     return isLauncherBooted;
+  },
+
+  checkIfLauncherNeedsToBeRevived: async function () {
+    const isLauncherRunning = config.getAdminPanelConfig('launcher.isRunning');
+    if (isLauncherRunning) {
+      this.startServerLauncher();
+    }
   },
 
   /**
@@ -256,41 +322,56 @@ module.exports = {
 
     let startProcessString = '';
 
-    // TODO: Windows
-    if (process.platform === 'linux') {
+    /**
+     * PKG_EXECPATH needs to be blank due to https://github.com/vercel/pkg/issues/376
+     */
+    if (os.isLinux()) {
       startProcessString = util.format(
-        'while : ; do PKG_EXECPATH=; nohup %s server-launcher %s && sleep 1 ; done &',
-        //'PKG_EXECPATH=; nohup %s server-launcher %s &',
+        // 'while true; do nohup PKG_EXECPATH=; %s server-launcher %s 1>/dev/null 2>/dev/null && sleep 1 ; done &',
+        'export PKG_EXECPATH=; export DEBUG=eqemu-admin:*; %s server-launcher %s &',
         pathManager.getEqemuAdminEntrypoint(),
         argString
       );
+
+      config.setAdminPanelConfig('launcher.isRunning', true);
+      console.log("[Process Manager] Starting server launcher...");
+
+      require('child_process').spawn(startProcessString,
+        {
+          cwd: pathManager.emuServerPath,
+          encoding: 'utf8',
+          shell: '/bin/bash',
+          detached: true
+        }
+      ).on('error', function (err) {
+        console.log(err);
+      });
+    }
+
+    if (os.isWindows()) {
+      if (pathManager.isRanFromPackagedNode()) {
+        startProcessString = util.format(
+          'set PKG_EXECPATH= && start /b %s server-launcher %s',
+          pathManager.getEqemuAdminEntrypoint(),
+          argString
+        );
+      } else if (pathManager.isRanAsStandaloneNodeProject()) {
+        startProcessString = util.format(
+          'set PKG_EXECPATH= && start /b node %s server-launcher %s',
+          pathManager.getEqemuAdminEntrypoint(),
+          argString
+        );
+      }
+
+      require('child_process').exec(startProcessString);
+
+      config.setAdminPanelConfig('launcher.isRunning', true);
+      console.log("[Process Manager] Starting server launcher...");
     }
 
     this.purgeServerLogs();
 
     debug('[startServerLauncher] Start string [%s]', startProcessString);
-
-    /**
-     * Start launcher
-     */
-    exec(startProcessString, { cwd: pathManager.emuServerPath });
-
-    /**
-     exec(startProcessString,
-     {
-        encoding: 'utf8',
-        // cwd: path.join(path.dirname(process.argv[0]), '../')
-      },
-     (error, stdout, stderr) => {
-        if (error) {
-          debug(`exec error: ${error}`);
-          return;
-        }
-        debug(`stdout: ${stdout}`);
-        debug(`stderr: ${stderr}`);
-      }
-     );
-     **/
 
     return this;
   },
@@ -301,7 +382,7 @@ module.exports = {
   async cancelRestart(options = []) {
     if (options.cancel) {
       this.cancelTimedRestart = true;
-      await serverDataService.messageWorld("[SERVER MESSAGE] Server restart cancelled")
+      await serverDataService.messageWorld('[SERVER MESSAGE] Server restart cancelled');
     }
   },
 
@@ -309,6 +390,7 @@ module.exports = {
    * @returns {exports}
    */
   restartServer: async function (options = []) {
+    this.init(options, true);
 
     /**
      * Delayed restart
@@ -332,9 +414,9 @@ module.exports = {
             Number((minutesRemaining).toFixed(2))
           );
 
-          await serverDataService.messageWorld(worldMessage)
+          await serverDataService.messageWorld(worldMessage);
 
-          debug('Sending world message | %s', worldMessage)
+          debug('Sending world message | %s', worldMessage);
         }
 
         if (unixTimeNow > endTime) {
@@ -370,7 +452,7 @@ module.exports = {
    */
   sleep: function (ms) {
     return new Promise(resolve => {
-      setTimeout(resolve, ms)
+      setTimeout(resolve, ms);
     });
   },
 
@@ -406,19 +488,38 @@ module.exports = {
       argString += arg + ' ';
     });
 
-    let isWindows          = process.platform === 'win32';
-    let isLinux            = process.platform === 'linux';
     let startProcessString = '';
-    if (isWindows) {
-      startProcessString = process_name + '.exe';
+    if (os.isWindows()) {
+      startProcessString =
+        util.format('bin\\%s',
+          process_name + '.exe'
+        );
+
+      require('child_process').spawn(startProcessString,
+        {
+          cwd: pathManager.emuServerPath,
+          detached: true,
+          stdio: 'ignore',
+          windowsHide: false
+        }
+      );
     }
 
-    if (isLinux) {
+    if (os.isLinux()) {
       startProcessString =
-        util.format('./bin/%s %s &',
+        util.format('nohup ./bin/%s %s 1>/dev/null 2>/dev/null &',
           process_name,
           argString
         );
+
+      require('child_process').spawn(startProcessString,
+        {
+          cwd: pathManager.emuServerPath,
+          encoding: 'utf8',
+          shell: '/bin/bash',
+          detached: true
+        }
+      );
     }
 
     debug('Starting process [%s] command [%s] path [%s]',
@@ -426,8 +527,6 @@ module.exports = {
       startProcessString,
       pathManager.getEmuServerPath()
     );
-
-    exec(startProcessString, { cwd: pathManager.emuServerPath, encoding: 'utf8' });
 
     this.processCount[process_name]++;
   },
@@ -437,7 +536,12 @@ module.exports = {
    */
   killProcess: function (process_id) {
     try {
-      require('child_process').execSync('kill -9 ' + process_id);
+      if (os.isLinux()) {
+        require('child_process').execSync('kill -9 ' + process_id);
+      }
+      if (os.isWindows()) {
+        require('child_process').execSync(util.format('TASKKILL /PID %s /F', process_id));
+      }
     } catch (e) {
       console.log(e);
     }
@@ -447,7 +551,7 @@ module.exports = {
    * @returns {Promise<void>}
    */
   pollProcessList: async function () {
-    this.systemProcessList = await psList();
+    this.systemProcessList = await this.getProcessList();
 
     let self = this;
     this.serverProcessNames.forEach(function (process_name) {
@@ -459,6 +563,47 @@ module.exports = {
         self.processCount[process.name]++;
       }
     });
+  },
+
+  async getProcessList() {
+    let processList = [];
+
+    if (os.isLinux()) {
+      processList = await psList();
+    }
+
+    if (os.isWindows()) {
+      const stdout = require('child_process')
+        .execSync("WMIC path win32_process get Description,Commandline,Processid /format:csv")
+        .toString();
+
+      stdout.split('\n').forEach((row) => {
+        if (row.includes(",") && !row.includes("Description,ProcessId")) {
+          if (row.split(",").length > 2) {
+            const splitRow          = row.split(",");
+            const splitLength       = splitRow.length;
+            const processId         = splitRow[splitLength - 1].trim();
+            const simpleProcessName = splitRow[splitLength - 2].replace('.exe', '').trim();
+            const commandLine       = row
+              .replace("," + splitRow[splitLength - 1].trim(), '')
+              .replace("," + splitRow[splitLength - 2].trim(), '')
+              .split(",")[1].trim()
+            ;
+
+            processList.push(
+              {
+                "name": simpleProcessName,
+                "pid": processId,
+                "cmd": commandLine
+              }
+            );
+          }
+        }
+      });
+    }
+
+    return processList;
+
   },
 
   /**
@@ -503,7 +648,7 @@ module.exports = {
       }
     });
 
-    debug('[isProcessRunning] found process [%s] [%s]', processName, foundProcess)
+    debug('[isProcessRunning] found process [%s] [%s]', processName, foundProcess);
 
     return foundProcess;
   },
@@ -512,8 +657,8 @@ module.exports = {
    * Purges server logs
    */
   purgeServerLogs() {
-    exec('rm -rf logs/zone/*.log', { cwd: pathManager.emuServerPath });
-    exec('rm -rf logs/*.log', { cwd: pathManager.emuServerPath });
+    exec('rm -rf logs/zone/*.log', {cwd: pathManager.emuServerPath});
+    exec('rm -rf logs/*.log', {cwd: pathManager.emuServerPath});
   },
 
   /**
@@ -546,7 +691,7 @@ module.exports = {
           fs.writeFileSync(logFile, data);
         }
 
-        process.exit()
+        process.exit();
       }
 
     }, WATCHDOG_TIMER);