+++ app/routes/api/server.js
@@ -4,10 +4,13 @@
  */
 const express              = require('express');
 const router               = express.Router();
-const serverProcessManager = require('../../core/server-process-manager')
-const config               = require('../../core/eqemu-config-service')
-const os                   = require('os')
+const serverProcessManager = require('../../core/server-process-manager');
+const config               = require('../../core/eqemu-config-service');
+const os                   = require('os');
 const si                   = require('systeminformation');
+let database               = use('/app/core/database');
+let dataService            = use('/app/core/eqemu-data-service-client.js');
+
 
 router.get('/hello', function (req, res, next) {
   res.send({ 'data': 'hello' });
@@ -19,26 +22,50 @@ router.get('/websocket-authorization', async function (req, res, next) {
   res.json(authorizedUser);
 });
 
+let lastCachedTime = 0;
+let lastCpuInfo = {};
+
+async function getCpuInfo() {
+  const currentTime = Math.floor(new Date() / 1000);
+
+  cpuInfo = lastCpuInfo;
+  if (currentTime - 30 > lastCachedTime || Object.keys(lastCpuInfo).length === 0) {
+    cpuInfo = {
+      info: await si.cpu(),
+      speed: await si.cpuCurrentspeed()
+    };
+    lastCpuInfo = cpuInfo;
+    lastCachedTime = currentTime;
+  }
+
+  cpuInfo.load = await si.currentLoad();
+
+  return cpuInfo;
+}
+
 router.get('/sysinfo', async function (req, res, next) {
+  const osMeta = use('/app/core/operating-system');
+  let disk     = {};
+
+  if (osMeta.isLinux()) {
+    disk = {
+      'io': await si.disksIO(),
+      'fs': {
+        'size': await si.fsSize(),
+        'stats': await si.fsStats()
+      }
+    };
+  }
+
   res.send(
     {
       'hostname': os.hostname(),
       'uptime': os.uptime(),
-      'cpu': {
-        info: await si.cpu(),
-        speed: await si.cpuCurrentspeed(),
-        load: await si.currentLoad()
-      },
+      'cpu': await getCpuInfo(),
       'mem': await si.mem(),
       'system': await si.system(),
       'os': await si.osInfo(),
-      'disk': {
-        'io': await si.disksIO(),
-        'fs': {
-          'size': await si.fsSize(),
-          'stats': await si.fsStats()
-        }
-      },
+      'disk': disk,
       'network': {
         'interfaces': await si.networkInterfaces(),
         'stats': await si.networkStats()
@@ -88,6 +115,34 @@ router.get('/log_categories', async function (req, res, next) {
   res.send(
     await models['logsys_categories'].resource.findAll({ order: [['log_category_description', 'ASC']] })
   );
+});
+
+router.get('/meta', async function (req, res, next) {
+  const ruleValues       = use('/app/repositories/ruleValuesRepository')
+  const currentExpansion = (await ruleValues.getCurrentExpansion());
+  const playersOnline    = (await dataService.getClientList());
+
+  const eqemu_config     = config.getServerConfig();
+  let meta               = {};
+  meta.long_name         = eqemu_config.server.world.longname;
+  meta.short_name        = eqemu_config.server.world.shortname;
+  meta.stats             = {};
+  meta.stats.accounts    = await database.tableRowCount(database.connection, 'account');
+  meta.stats.characters  = await database.tableRowCount(database.connection, 'character_data');
+  meta.stats.guilds      = await database.tableRowCount(database.connection, 'guilds');
+  meta.stats.items       = await database.tableRowCount(database.contentConnection, 'items');
+  meta.stats.npcs        = await database.tableRowCount(database.contentConnection, 'npc_types');
+  meta.zone_count        = (await dataService.getZoneList()).length;
+  meta.players_online    = (playersOnline) ? playersOnline.length : 0;
+  meta.current_expansion = (currentExpansion) ? currentExpansion : 0;
+  meta.process_counts    = await serverProcessManager.getProcessCounts();
+  meta.uptime            = await dataService.getWorldUptime();
+
+  res.json(meta);
+});
+
+router.get('/schema', async function (req, res, next) {
+  res.json((await dataService.getDatabaseSchema()));
 });
 
 module.exports = router;